ai-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
openai==1.12.0
langchain==0.1.7
langchain-openai==0.0.6
langchain-community==0.0.20
google-cloud-vision==3.7.1
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
python-multipart==0.0.9
httpx==0.26.0
redis[asyncio]==5.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
faiss-cpu==1.7.4
tiktoken==0.6.0
aiofiles==23.2.1
boto3==1.34.44
python-jose==3.3.0
cryptography==42.0.3
